refactor: 調整專案結構，將視窗分層至 Views 並抽出 Layout
