+++ src/OverTranslate/App.xaml.cs
@@ -5,6 +5,9 @@ using OverTranslate.Services;
 using OverTranslate.Models;
 using System.Threading.Tasks;
 using Velopack;
+using OverTranslate.Views;
+using OverTranslate.Views.Shell;
+using OverTranslate.Views.Translation;
 
 namespace OverTranslate;
