--- src/OverTranslate/AssemblyInfo.cs
@@ -1,12 +1,0 @@
-using System.Runtime.CompilerServices;
-using System.Windows;
-
-[assembly: InternalsVisibleTo("OverTranslate.Tests")]
-[assembly: ThemeInfo(
-    ResourceDictionaryLocation.None,            //where theme specific resource dictionaries are located
-                                                //(used if a resource is not found in the page,
-                                                // or application resource dictionaries)
-    ResourceDictionaryLocation.SourceAssembly   //where the generic resource dictionary is located
-                                                //(used if a resource is not found in the page,
-                                                // app, or any theme specific resource dictionaries)
-)]
