+++ src/OverTranslate/Views/Capture/ScreenCaptureWindow.xaml
@@ -1,4 +1,4 @@
-<Window x:Class="OverTranslate.ScreenCaptureWindow"
+<Window x:Class="OverTranslate.Views.Capture.ScreenCaptureWindow"
         xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
         xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
         WindowStyle="None"
