+++ src/OverTranslate/Views/Capture/ToolbarWindow.xaml
@@ -1,4 +1,4 @@
-<Window x:Class="OverTranslate.ToolbarWindow"
+<Window x:Class="OverTranslate.Views.Capture.ToolbarWindow"
         xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
         xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
         WindowStyle="None"
