+++ src/OverTranslate/Views/MainWindow.xaml
@@ -1,4 +1,4 @@
-<Window x:Class="OverTranslate.MainWindow"
+<Window x:Class="OverTranslate.Views.MainWindow"
         xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
         xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
         Title="OverTranslate"
