+++ src/OverTranslate/Views/Settings/SettingsWindow.xaml
@@ -1,4 +1,4 @@
-<Window x:Class="OverTranslate.SettingsWindow"
+<Window x:Class="OverTranslate.Views.Settings.SettingsWindow"
         xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
         xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
         Title="OverTranslate 設定"
