+++ src/OverTranslate/Views/Shell/ToastWindow.xaml
@@ -1,4 +1,4 @@
-<Window x:Class="OverTranslate.ToastWindow"
+<Window x:Class="OverTranslate.Views.Shell.ToastWindow"
         xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
         xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
         WindowStyle="None"
