+++ src/OverTranslate/Views/Shell/ToastWindow.xaml.cs
@@ -1,7 +1,7 @@
 using System.Windows;
 using System.Windows.Media.Animation;
 
-namespace OverTranslate;
+namespace OverTranslate.Views.Shell;
 
 public partial class ToastWindow : Window
 {
