+++ src/OverTranslate/Views/Shell/TrayMenuWindow.xaml
@@ -1,4 +1,4 @@
-<Window x:Class="OverTranslate.TrayMenuWindow"
+<Window x:Class="OverTranslate.Views.Shell.TrayMenuWindow"
         xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
         xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
         WindowStyle="None"
