+++ src/OverTranslate/Views/Shell/UpdateWindow.xaml
@@ -1,4 +1,4 @@
-<Window x:Class="OverTranslate.UpdateWindow"
+<Window x:Class="OverTranslate.Views.Shell.UpdateWindow"
         xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
         xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
         Title="發現新版本"
